+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.10-dev0
+
+### Enhancements
+
+* **Migrate Jira Source connector from V1 to V2**
+* **Add Jira Source connector integration and unit tests**
+
 ## 0.5.9
 
 ### Features